Update README.md to use gradlew rather than gradle
`./gradlew` will use whatever version of Gradle is installed to locally download the version of Gradle corresponding with this build system. The version of Gradle from Debian and Ubuntu's packages is severely out of date compared with the Graph Extractor, so using the wrapper will make it possible to run on old versions of Gradle.