Inherit CDFG from the EnhancedUnitGraph
Now the CDGF inherit from the EnhancedUnitGraph insted that from the
BriefUnitGraph.
We have to figure out if this is a reasonable choice.

The change is due to the fact that we always build the
EnhancedUnitGraph for the sliced body.
This has side effect on the body itself, leading to possible issues.