Explicitly set the sbt version
The oneJar plugin does not work with newer versions of sbt (> 1.x I think, I tried it with 1.1.4).

Thus, we have to set manually a sbt version to use, and 0.13.17 works.
Note that any version below of 0.13 has a bug raising the exception "Error wrapping InputStream in GZIPInputStream: java.io.EOFException"