Fix (again) the issue of using apkinfo.jar
apkinfo.jar is stored as a resource in the jar file of the graph extractor. However, we cannot access that as a "file" directly, as we would do on a "normal" jar.

The issue is that when using oneJar the resources file gets packaged differently. A quick solution is to save the jar file on disk before running that.
This solution works both if we're using sbt, a "standard" jar or a oneJar (as we do when we deploy the extractor).